Move "ClosureRunCommand" to "UserInterface\Cli" namespace
